% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ats.R
\name{add_ats}
\alias{add_ats}
\title{Add attributes to data object}
\usage{
add_ats(.x, ...)
}
\arguments{
\item{.x}{Data object on which to add attributes.}

\item{...}{Named elements to be attached to .x as attributes.}
}
\description{
Add attributes to data object
}
