% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr.R
\name{add_attr}
\alias{add_attr}
\title{Add attributes to data object}
\usage{
add_attr(.x, ...)
}
\arguments{
\item{.x}{Data object on which to add attributes.}

\item{...}{Named elements to be attached to .x as attributes.}
}
\description{
Add attributes to data object
}
