% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atselect.R
\docType{methods}
\name{atselect,list-method}
\alias{atselect,list-method}
\title{An S4 class to represent an atselect method}
\usage{
\S4method{atselect}{list}(.x, ...)
}
\description{
An S4 class to represent an atselect method
}
