% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atselect.R
\name{atselect}
\alias{atselect}
\title{select attributes}
\usage{
atselect(.x, ...)
}
\arguments{
\item{.x}{Data object containing attributes}

\item{...}{Name of attributes to select}
}
\value{
Selected attributes
}
\description{
select attributes
}
